+++ doc/transform_methods/comparison/fig_experiment/experiment.py
@@ -54,9 +54,6 @@ for num, (ax, (label, transFunc, color), letter) in enumerate(zip(axs.ravel(),
     else:
         trans = transFunc(Q0, direction="inverse", **targs)
 
-    if label == 'linbasex':  # bugfix smoothing=0 transform offset by 1 pixel
-        trans[:, 1:] = trans[:, :-1]
-
     r, inten = abel.tools.vmi.angular_integration_3D(trans[::-1],
                                                      origin=(0, 0),
                                                      dr=0.1)